requirements.txt
Flask==0.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.19
blinker==1.3
itsdangerous==0.24
kombu==3.0.30
pytz==2014.10
redis==2.10.5
cloudinary==1.5.0
gunicorn==19.4.5
pdfminer.six